+++ tests/test_serialize_fmu_state.py
@@ -1,6 +1,5 @@
 import unittest
-from unittest import skipIf
-from fmpy import read_model_description, extract, platform
+from fmpy import read_model_description, extract
 from fmpy.fmi2 import FMU2Slave
 from fmpy.util import download_test_file
 import shutil
@@ -8,7 +7,6 @@ import shutil
 
 class SerializeFMUStateTest(unittest.TestCase):
 
-    @skipIf(platform == 'linux64', "Crashes on Linux")
     def test_serialize_fmu_state(self):
 
         fmu_filename = 'Rectifier.fmu'
@@ -39,14 +37,14 @@ class SerializeFMUStateTest(unittest.TestCase):
         # serialize the FMU state
         serialized_state = fmu.serializeFMUstate(state)
 
-        # de-serialize the FMU state
-        fmu.deSerializeFMUstate(serialized_state, state)
+        # de-serialize the FMU state (re-using memory)
+        deserialized_state = fmu.deSerializeFMUstate(serialized_state, state)
 
         # set the FMU state
-        fmu.setFMUstate(state)
+        fmu.setFMUstate(deserialized_state)
 
         # free the FMU state
-        fmu.freeFMUstate(state)
+        fmu.freeFMUstate(deserialized_state)
 
         fmu.terminate()
         fmu.freeInstance()